refactor: extract OHLCVDataProviderBase to eliminate duplication
Introduce an intermediate base class OHLCVDataProviderBase that handles
all shared OHLCV data provider logic:
- Constructor, copy(), get_data(), prepare_backtest_data(), get_backtest_data()
- CSV storage helpers, date range resolution, number of data points
- Fixes bug: create_start_date was missing from Yahoo/AV/Polygon providers

Concrete providers now only implement:
- market_name, timeframe_map (class attributes)
- _download_ohlcv() (API-specific download)
- _validate_symbol() (optional, Yahoo only)

Reduces Yahoo/AlphaVantage/Polygon from ~1,851 lines to ~1,033 total (44% reduction).
CCXT intentionally excluded (different pattern: exchange-based detection, pagination, ticker support).
