+++ tests/infrastructure/data_providers/test_alpha_vantage_ohlcv_data_provider.py
@@ -228,13 +228,13 @@ class TestAlphaVantageTimeFrame(TestCase):
     def test_supported_timeframes(self):
         for tf in ["1m", "5m", "15m", "30m", "1h", "1d", "1W", "1M"]:
             provider = AlphaVantageOHLCVDataProvider(time_frame=tf)
-            config = provider._get_av_config()
+            config = provider._get_provider_interval()
             self.assertIsNotNone(config)
 
     def test_unsupported_timeframe_raises(self):
         provider = AlphaVantageOHLCVDataProvider(time_frame="2m")
         with self.assertRaises(Exception) as cm:
-            provider._get_av_config()
+            provider._get_provider_interval()
         self.assertIn("not supported", str(cm.exception))
